moved shulker dye functionality from the mixin against BlockBehavior.useItemOn to a mixin against Item.useOn.
reasoning: In Minecraft, if a block has its own functionality, it should take priority when the player clicks the 'use' button regardless of what they are holding. Clicking the 'use' button while crouching shifts the priority to the item held. Setting this functionality against BlockBehavior.useItemOn was overriding the shulker box block's functionality making it so the player could not open the box if they were holding dye. Moving the functionality to Item.useOn causes it to be triggered when the player is crouching.

attempting to dye a shulker its current color now does nothing.
successfully dying a shulker now consumes the dye & the dye's usage statistics is updated.