A number of changes to modify the GUI's new caching behavior for rendering items to account for shulker boxes having lid items.
MixinSpecialModelWrapper.update added to include lidItem data to the hash used to by the gui renderer to identify unique items in the texture atlas. This allows each combination of shulker/lidItem/scale to be rendered independently in the GUI.
MixinLayerRenderState.setupShulkerBoxSpecialRenderer and
MixinLayerRenderState.render added to pass the custom data from the setup pass to the render pass.
IShulkerRendererLidItem introduced as an interface to facilitate pulling and setting custom data on for the above modifications.

update to version 0.4.0 of permissions api